requirements.txt
pywinauto
pyyaml
pyautogui
opencv-python
pytesseract
wmi
pywin32
numpy
Pillow
winaudio
requests
schema
pydantic
colour
ultralytics
selenium
mss
